refactoring: move class: filter.Cleanse -> stream.processor.Cleanse
