+++ src/main/java/org/bricolages/streaming/filter/FloatOp.java
@@ -1,4 +1,5 @@
 package org.bricolages.streaming.filter;
+import org.bricolages.streaming.stream.processor.Cleanse;
 import lombok.*;
 
 public class FloatOp extends SingleColumnOp {
