+++ src/main/java/org/bricolages/streaming/filter/IntOp.java
@@ -1,4 +1,5 @@
 package org.bricolages.streaming.filter;
+import org.bricolages.streaming.stream.processor.Cleanse;
 import lombok.*;
 
 class IntOp extends SingleColumnOp {
