+++ src/main/java/org/bricolages/streaming/stream/processor/BooleanColumnProcessor.java
@@ -1,5 +1,5 @@
 package org.bricolages.streaming.stream.processor;
-import org.bricolages.streaming.filter.*;
+import org.bricolages.streaming.filter.FilterException;
 import java.util.Objects;
 import lombok.*;
