+++ src/main/java/org/bricolages/streaming/stream/processor/DateColumnProcessor.java
@@ -1,5 +1,5 @@
 package org.bricolages.streaming.stream.processor;
-import org.bricolages.streaming.filter.*;
+import org.bricolages.streaming.filter.FilterException;
 import org.bricolages.streaming.exception.*;
 import java.time.ZoneOffset;
 import java.time.OffsetDateTime;
