+++ src/main/java/org/bricolages/streaming/stream/processor/SmallintColumnProcessor.java
@@ -1,5 +1,5 @@
 package org.bricolages.streaming.stream.processor;
-import org.bricolages.streaming.filter.*;
+import org.bricolages.streaming.filter.FilterException;
 import lombok.*;
 
 public class SmallintColumnProcessor extends SingleColumnProcessor {
